DiscountPolicy의 인터페이스화, 트레이드오프
