+++ chapter2/src/main/java/com/example/chapter2/src/discount/AmountDiscountPolicy.java
@@ -3,9 +3,7 @@ package com.example.chapter2.src.discount;
 import com.example.chapter2.src.Money;
 import com.example.chapter2.src.Screening;
 
-import java.util.List;
-
-public class AmountDiscountPolicy extends DiscountPolicy{
+public class AmountDiscountPolicy extends DefaultDiscountPolicy {
     private Money discountAmount;
 
     public AmountDiscountPolicy(Money discountAmount,DiscountCondition ... conditions) {
