+++ chapter2/src/main/java/com/example/chapter2/src/discount/DefaultDiscountPolicy.java
@@ -1,0 +1,28 @@
+package com.example.chapter2.src.discount;
+
+import com.example.chapter2.src.Money;
+import com.example.chapter2.src.Screening;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+
+public abstract class DefaultDiscountPolicy implements DiscountPolicy{
+    private List<DiscountCondition> conditions = new ArrayList<>();
+
+    public DefaultDiscountPolicy(DiscountCondition ... conditions) {
+        this.conditions = Arrays.asList(conditions);
+    }
+
+    @Override
+    public Money calculateDiscountAmount(Screening screening){
+        for(DiscountCondition each: conditions){
+            if(each.isSatisfiedBy(screening)){
+                return getDiscountAmount(screening);
+            }
+        }
+        return Money.ZERO;
+    }
+
+    abstract protected Money getDiscountAmount(Screening screening);
+}
