+++ chapter2/src/main/java/com/example/chapter2/src/discount/NoneDiscountPolicy.java
@@ -3,9 +3,10 @@ package com.example.chapter2.src.discount;
 import com.example.chapter2.src.Money;
 import com.example.chapter2.src.Screening;
 
-public class NoneDiscountPolicy extends DiscountPolicy{
+public class NoneDiscountPolicy implements DiscountPolicy {
+
     @Override
-    protected Money getDiscountAmount(Screening screening) {
-        return Money.ZERO;
+    public Money calculateDiscountAmount(Screening screening) {
+        return null;
     }
 }
